pyproject.toml
[project]
name = "dev"
version = "1.0.0"
description = "Face Similarity Search Platform - Find your photos using face recognition"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Aditya", email = "aditya@example.com" }
]
keywords = ["face-recognition", "similarity-search", "faiss", "insightface"]

dependencies = [
    # FastAPI & Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    # AI/ML
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "Pillow>=10.2.0",
    "qrcode>=7.4.0",
    # Logging
    "structlog>=24.1.0",
    "opencv-python>=4.13.0.90",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tell hatchling where the package is
[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
    "ipython>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
